converted 32 bits integers to 64 bits, ensure consistency

+++ libgramtools/include/bwt_search.hpp
@@ -33,7 +33,7 @@ using sequence_set = std::unordered_set<SEQUENCE, seq_hash<SEQUENCE>>;
 void calc_kmer_matches(const FM_Index &fm_index, int k,
                        sequence_map<std::vector<uint8_t>, std::list<std::pair<uint64_t, uint64_t>>> &kmer_idx,
                        sequence_map<std::vector<uint8_t>, std::list<std::pair<uint64_t, uint64_t>>> &kmer_idx_rev,
-                       sequence_map<std::vector<uint8_t>, std::list<std::vector<std::pair<uint32_t, std::vector<int>>>>> &kmer_sites,
+                       sequence_map<std::vector<uint8_t>, std::list<std::vector<std::pair<uint64_t, std::vector<int>>>>> &kmer_sites,
                        std::vector<int> &mask_a, uint64_t maxx, sequence_set<std::vector<uint8_t>> &kmers_in_ref,
                        std::vector<std::vector<uint8_t>> &kmerfile,
                        const DNA_Rank &rank_all);
@@ -41,8 +41,8 @@ void calc_kmer_matches(const FM_Index &fm_index, int k,
 void precalc_ranks(const FM_Index &fm_index, const DNA_Rank& rank_all);
 
 std::pair<uint64_t, uint64_t> bidir_search(const uint8_t next_char,
-                                           const std::list<std::pair<unsigned long, unsigned long>>::iterator &sa_interval_it,
-                                           const std::list<std::pair<unsigned long, unsigned long>>::iterator &sa_interval_it_rev,
+                                           const std::list<std::pair<uint64_t, uint64_t>>::iterator &sa_interval_it,
+                                           const std::list<std::pair<uint64_t, uint64_t>>::iterator &sa_interval_it_rev,
                                            const DNA_Rank &rank_all, const FM_Index &fm_index);
 
 bool skip(uint64_t &left, uint64_t &right, uint64_t &left_rev, uint64_t &right_rev, const uint64_t maxx, const uint64_t num,
@@ -55,7 +55,7 @@ std::vector<uint8_t>::iterator bidir_search_fwd(const FM_Index &fm_index,
                                                 std::vector<uint8_t>::iterator pat_end,
                                                 std::list<std::pair<uint64_t, uint64_t>> &sa_intervals,
                                                 std::list<std::pair<uint64_t, uint64_t>> &sa_intervals_rev,
-                                                std::list<std::vector<std::pair<uint32_t, std::vector<int>>>> &sites,
+                                                std::list<std::vector<std::pair<uint64_t, std::vector<int>>>> &sites,
                                                 std::vector<int> &mask_a, uint64_t maxx, bool &first_del,
                                                 bool kmer_precalc_done, const DNA_Rank &rank_all);
